Whitelist SIP providers based on their endpoint IP address
This script is intended to reduce traffic on udp:5060 to an Asterisk instance. Run voip_iptables.py and then add a corresponding iptables INPUT rule, somewhere convenient, to direct traffic to the voip_inbound chain. It's pretty rough going but I needed to alleviate the load.

The allowed_hosts.txt file contains the currently known hostnames for voip.ms and Unlimitel endpoints.